Extract terminal fullscreen into composed FullscreenController
Move the 8 fullscreen methods and all fullscreen state out of TerminalWidget
into a new terminal_fullscreen.py FullscreenController(terminal) that the widget
owns as self._fullscreen. Behavior-preserving; TerminalWidget keeps a thin
toggle_fullscreen forwarder and calls self._fullscreen.setup_shortcut() at init.

The two identical restore_focus closures (enter/exit) are consolidated into one
_restore_focus method (the only non-move change). Characterization tests updated
to target the controller, plus a restore_focus fallback-tier test.
