+++ tests/test_terminal_fullscreen.py
@@ -9,11 +9,13 @@ tested as a method once extracted (see test_restore_focus_* below).
 Standard headless pattern: build a bare TerminalWidget via ``__new__`` and stub
 only the attributes each method touches.
 """
+from types import SimpleNamespace
+
 import pytest
 
 pytest.importorskip("gi")
 
-from sshpilot.terminal import TerminalWidget
+from sshpilot.terminal_fullscreen import FullscreenController
 
 
 class _Container:
@@ -28,48 +30,68 @@ class _Container:
         self.visible = v
 
 
-def _tw(**attrs):
-    tw = TerminalWidget.__new__(TerminalWidget)
-    tw._is_fullscreen = False
-    tw._fullscreen_banner_container = None
+def _fc(**attrs):
+    fc = FullscreenController.__new__(FullscreenController)
+    fc.t = SimpleNamespace()
+    fc._is_fullscreen = False
+    fc._fullscreen_banner_container = None
     for k, v in attrs.items():
-        setattr(tw, k, v)
-    return tw
+        setattr(fc, k, v)
+    return fc
 
 
 def test_toggle_enters_when_not_fullscreen():
-    tw = _tw(_is_fullscreen=False)
+    fc = _fc(_is_fullscreen=False)
     calls = []
-    tw._enter_fullscreen = lambda: calls.append("enter")
-    tw._exit_fullscreen = lambda: calls.append("exit")
-    tw.toggle_fullscreen()
+    fc._enter_fullscreen = lambda: calls.append("enter")
+    fc._exit_fullscreen = lambda: calls.append("exit")
+    fc.toggle_fullscreen()
     assert calls == ["enter"]
 
 
 def test_toggle_exits_when_fullscreen():
-    tw = _tw(_is_fullscreen=True)
+    fc = _fc(_is_fullscreen=True)
     calls = []
-    tw._enter_fullscreen = lambda: calls.append("enter")
-    tw._exit_fullscreen = lambda: calls.append("exit")
-    tw.toggle_fullscreen()
+    fc._enter_fullscreen = lambda: calls.append("enter")
+    fc._exit_fullscreen = lambda: calls.append("exit")
+    fc.toggle_fullscreen()
     assert calls == ["exit"]
 
 
 def test_show_fullscreen_banner_makes_it_visible():
-    tw = _tw(_fullscreen_banner_container=_Container(visible=False))
-    tw._show_fullscreen_banner()
-    assert tw._fullscreen_banner_container.visible is True
+    fc = _fc(_fullscreen_banner_container=_Container(visible=False))
+    fc._show_fullscreen_banner()
+    assert fc._fullscreen_banner_container.visible is True
 
 
 def test_hide_fullscreen_banner_makes_it_invisible():
-    tw = _tw(_fullscreen_banner_container=_Container(visible=True))
-    tw._hide_fullscreen_banner()
-    assert tw._fullscreen_banner_container.visible is False
+    fc = _fc(_fullscreen_banner_container=_Container(visible=True))
+    fc._hide_fullscreen_banner()
+    assert fc._fullscreen_banner_container.visible is False
 
 
 def test_dismiss_hides_banner():
-    tw = _tw()
+    fc = _fc()
     hidden = []
-    tw._hide_fullscreen_banner = lambda: hidden.append(True)
-    tw._on_fullscreen_banner_dismiss(None)
+    fc._hide_fullscreen_banner = lambda: hidden.append(True)
+    fc._on_fullscreen_banner_dismiss(None)
     assert hidden == [True]
+
+
+def test_restore_focus_prefers_backend_then_vte_then_self():
+    # backend present → backend.grab_focus
+    grabbed = []
+    fc = _fc()
+    fc.t = SimpleNamespace(backend=SimpleNamespace(grab_focus=lambda: grabbed.append("backend")),
+                           vte=SimpleNamespace(grab_focus=lambda: grabbed.append("vte")),
+                           grab_focus=lambda: grabbed.append("self"))
+    assert fc._restore_focus() is False
+    assert grabbed == ["backend"]
+
+    # no backend → vte.grab_focus
+    grabbed.clear()
+    fc.t = SimpleNamespace(backend=None,
+                           vte=SimpleNamespace(grab_focus=lambda: grabbed.append("vte")),
+                           grab_focus=lambda: grabbed.append("self"))
+    fc._restore_focus()
+    assert grabbed == ["vte"]
